To-Do List
This is the Excel version of the diary, shared to the team via Google Drive. This is to keep track of specific tasks, so we can see exactly where we are and what we have left to do. It is to be updated as soon as tasks are complete, ideally with a semi-detailed description of what specifically has been done.